Add SimResult & adjust occbin returns
Introduce SimResult and OccBinDiagnostics (SymbolicDSGE/core/sim_result.py) and update SolvedModel.sim and .irf to return SimResult objects (include states, observables, regimes/diagnostics). Cast occbin_solve wire-format int8 outputs to int64 for regimes and periodic in the Cython kernel. Simplify occbin __init__ exports. Update UI session serialization and tests to import/handle the new return types and internal occbin symbols. These changes package simulation results and OccBin diagnostics into a stable dataclass API.

+++ SymbolicDSGE/_ckernels/occbin/_occbin.pyx
@@ -647,11 +647,12 @@ def occbin_solve(a, b, c, f_ref, ss, par, size_t cond_addr,
 
     if status != SDSGE_OCCBIN_PERIOD_OK:
         raise RuntimeError(_solve_error(status, &diag, max_iter))
-    return out, regimes, {
+    # int8 is the kernel's wire format; callers get a width that will not wrap.
+    return out, regimes.astype(np.int64), {
         "T_used": T_used,
         "iters": iters,
         "max_err": max_err,
-        "periodic": periodic,
+        "periodic": periodic.astype(np.int64),
     }
